# docs/api-spec.yaml
openapi: 3.0.0
paths:
...
  /categories:
    get:
      operationId: CategoriesController_list
      parameters:
        - name: raw
          required: false
          in: query
          schema:
            type: string
        - name: lang
          required: false
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags:
        - categories
    post:
      operationId: CategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - categories
  /categories/{id}:
    get:
      operationId: CategoriesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: raw
          required: false
          in: query
          schema:
            type: string
        - name: lang
          required: false
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags:
        - categories
    patch:
      operationId: CategoriesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - categories
    delete:
      operationId: CategoriesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - categories
  /products:
    get:
      operationId: ProductsController_list
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: categoryId
          required: false
          in: query
          description: Kategori ID filtresi
          schema:
            type: string
        - name: q
          required: false
          in: query
          description: Ad veya stok kodunda arama
          schema:
            type: string
        - name: sort
          required: false
          in: query
          schema:
            default: newest
            type: string
            enum:
              - newest
              - price_asc
              - price_desc
        - name: onlyDiscounted
          required: false
          in: query
          description: Sadece indirimli ürünler (original_price > price)
          schema:
            type: boolean
        - name: onlyNew
          required: false
          in: query
          description: Sadece yeni ürünler
          schema:
            type: boolean
        - name: inStock
          required: false
          in: query
          description: Sadece stokta olanlar
          schema:
            type: boolean
        - name: lang
          required: false
          in: query
          description: Dil kodu (tr/en/de/fr/ar/nl)
          schema:
            enum:
              - tr
              - en
              - de
              - fr
              - ar
              - nl
            type: string
        - name: raw
          required: false
          in: query
          description: true → ham çok dilli map
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - products
    post:
      operationId: ProductsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - products
  /products/{id}:
    get:
      operationId: ProductsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: raw
          required: false
          in: query
          schema:
            type: string
        - name: lang
          required: false
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags:
        - products
    patch:
      operationId: ProductsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - products
    delete:
      operationId: ProductsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - products
  /products/{id}/similar:
    get:
      operationId: ProductsController_listSimilar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags:
        - products
  /products/{id}/images:
    post:
      operationId: ProductsController_addImage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProductImageDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - products
  /products/{id}/images/{imageId}:
    delete:
      operationId: ProductsController_removeImage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: imageId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - products
  /cart:
    get:
      operationId: CartController_list
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - cart
    delete:
      operationId: CartController_clear
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - cart
  /cart/items:
    post:
      operationId: CartController_addItem
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - cart
  /cart/items/{productId}:
    patch:
      operationId: CartController_updateItem
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - cart
    delete:
      operationId: CartController_removeItem
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - cart
  /cart/checkout-preview:
    post:
      operationId: CartController_checkoutPreview
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutPreviewDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - cart
...
  /announcements:
    get:
      operationId: AnnouncementsController_list
      parameters:
        - name: raw
          required: false
          in: query
          schema:
            type: string
        - name: lang
          required: false
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags:
        - announcements
    post:
      operationId: AnnouncementsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnnouncementDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - announcements
  /announcements/{id}:
    get:
      operationId: AnnouncementsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: raw
          required: false
          in: query
          schema:
            type: string
        - name: lang
          required: false
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags:
        - announcements
    patch:
      operationId: AnnouncementsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnnouncementDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - announcements
    delete:
      operationId: AnnouncementsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - announcements
...
info:
  title: MarketApp API
  description: MarketApp backend API dokümantasyonu
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: object
          description: Çok dilli kategori adı — tr zorunlu
          example:
            tr: Atıştırmalık
            en: Snacks
        image_url:
          type: string
        display_order:
          type: number
          default: 0
      required:
        - name
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: object
          description: Çok dilli kategori adı — tr zorunlu
          example:
            tr: Atıştırmalık
            en: Snacks
        image_url:
          type: string
        display_order:
          type: number
          default: 0
    CreateProductDto:
      type: object
      properties:
        category_id:
          type: string
        name:
          type: object
          description: 'Çok dilli ürün adı, ör. { "tr": "...", "en": "..." } — tr zorunlu'
          example:
            tr: Koska Sade Tahin Helvası 200 Gr
        sku:
          type: string
          description: Stok kodu (SKU), benzersiz olmalı
        description:
          type: object
          description: 'Çok dilli açıklama, ör. { "tr": "...", "de": "..." }'
        price:
          type: number
        original_price:
          type: number
          description: Doluysa ve price'tan büyükse ürün indirimli sayılır
        is_new_arrival:
          type: boolean
          default: false
        stock_quantity:
          type: number
          default: 0
        images:
          description: Başlangıç ürün görseli URL listesi
          type: array
          items:
            type: string
      required:
        - category_id
        - name
        - sku
        - price
    UpdateProductDto:
      type: object
      properties:
        category_id:
          type: string
        name:
          type: object
          description: 'Çok dilli ürün adı, ör. { "tr": "...", "en": "..." } — tr zorunlu'
          example:
            tr: Koska Sade Tahin Helvası 200 Gr
        sku:
          type: string
          description: Stok kodu (SKU), benzersiz olmalı
        description:
          type: object
          description: 'Çok dilli açıklama, ör. { "tr": "...", "de": "..." }'
        price:
          type: number
        original_price:
          type: number
          description: Doluysa ve price'tan büyükse ürün indirimli sayılır
        is_new_arrival:
          type: boolean
          default: false
        stock_quantity:
          type: number
          default: 0
    AddProductImageDto:
      type: object
      properties:
        image_url:
          type: string
        display_order:
          type: number
          default: 0
      required:
        - image_url
    AddCartItemDto:
      type: object
      properties:
        product_id:
          type: string
        quantity:
          type: number
          default: 1
      required:
        - product_id
    UpdateCartItemDto:
      type: object
      properties:
        quantity:
          type: number
      required:
        - quantity
    CheckoutPreviewDto:
      type: object
      properties:
        coupon_code:
          type: string
          description: Uygulanacak kupon kodu (opsiyonel)
...
    CreateAnnouncementDto:
      type: object
      properties:
        title:
          type: object
          description: 'Çok dilli başlık, ör. { "tr": "...", "en": "..." }'
        content:
          type: object
          description: Çok dilli içerik — tr zorunlu
          example:
            tr: 'Kupon Kodu: ilksiparis10'
        image_url:
          type: string
      required:
        - content
    UpdateAnnouncementDto:
      type: object
      properties:
        title:
          type: object
          description: 'Çok dilli başlık, ör. { "tr": "...", "en": "..." }'
        content:
          type: object
          description: Çok dilli içerik — tr zorunlu
          example:
            tr: 'Kupon Kodu: ilksiparis10'
        image_url:
          type: string
...